Move time slices tests to platform
